REFACTOR: Replace Client Model #28
