+++ config/urls.py
@@ -28,7 +28,7 @@ router = DefaultRouter()
 router.register(r'categories', CategoryViewSet)
 router.register(r'telephones', TelephoneViewSet)
 router.register(r'addresses', AddressViewSet)
-router.register(r'users/register/clients', ClientViewSet)
+router.register(r'users/register/naturalPerson', NaturalPersonViewSet)
 router.register(r'users/register/restaurants', RestaurantViewSet)
 
 urlpatterns = [
