+++ core/admin.py
@@ -5,5 +5,5 @@ from core.models import *
 admin.site.register(Category)
 admin.site.register(Address)
 admin.site.register(Telephone)
-admin.site.register(Client)
+admin.site.register(NaturalPerson)
 admin.site.register(Restaurant)
