+++ core/models/__init__.py
@@ -1,5 +1,5 @@
 from .category import Category
 from .telephone import Telephone
 from .address import Address
-from .client import Client
+from .natural_person import NaturalPerson
 from .restaurant import Restaurant
