+++ core/serializers/__init__.py
@@ -2,5 +2,5 @@ from .user import UserRegisterSerializer
 from .category import CategorySerializer
 from .telephone import TelephoneSerializer
 from .address import AddressSerializer
-from .client import ClientSerializer
+from .natural_person import NaturalPersonSerializer
 from .restaurant import RestaurantSerializer
